Issue: WSW-25371 Port code to fix wlan_set_rf_tx_power is not consistent with tx-frame power output
[Issue]
wlan_set_rf_tx_power is not consistent with tx-frame power output.
Build error in wifi_set_rf_rx_antenna() for wifi_test_mode example: operand of "*" must be a pointer but has type "uint8_t"
[Fix]
We need to convert user power vals including -ve vals as per labtool. For e.g. if we configure power to be 10 dBm then FW should get the power level value as 160 since the resolution is 1/16.
Set rx_antenna_set to 1 when SUCCESS in wifi_set_rf_rx_antenna().
[Validation done]
1. Compilation successful
2. Run "wlan-set-rf-tx-power 10 0 0", check the content of related command, it was as expected.
3. Run "wlan-set-rf-rx-antenna 1", it's ok.
[Open Issues/Todo]
None
[Reviewers]
Gannicus Guo, Sandip Dalvi, Pan Gao, Maochen Wang, Hui Bai

+++ incl/wifidriver/wifi.h
@@ -1467,7 +1467,7 @@ int wifi_set_rf_rx_antenna(const uint8_t antenna);
 
 int wifi_get_rf_rx_antenna(uint8_t *antenna);
 
-int wifi_set_rf_tx_power(const uint8_t power, const uint8_t mod, const uint8_t path_id);
+int wifi_set_rf_tx_power(const uint32_t power, const uint8_t mod, const uint8_t path_id);
 
 int wifi_cfg_rf_he_tb_tx(uint16_t enable, uint16_t qnum, uint16_t aid, uint16_t axq_mu_timer, int16_t tx_power);
